backend/requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiohttp-cors
aiosignal
airportsdata
altair
annotated-types
anyio
astor
attrs
blake3
blinker
cachetools
certifi
charset-normalizer
click
cloudpickle
colorama
colorful
compressed-tensors
Deprecated
depyf
dill
diskcache
distlib
distro
einops
fastapi
filelock
frozenlist
fsspec
gguf
gitdb
GitPython
google-api-core
google-auth
googleapis-common-protos
grpcio
h11
httpcore
httptools
httpx
huggingface-hub
idna
importlib_metadata
interegular
Jinja2
jiter
joblib
jsonschema
jsonschema-specifications
lark
limits
lm-format-enforcer
lxml
markdown-it-py
MarkupSafe
mdurl
mistral_common
mpmath
msgpack
msgspec
multidict
narwhals
nest-asyncio
networkx
numpy
nvidia-ml-py
openai
opencensus
opencensus-context
opencv-python-headless
outlines
outlines_core
packaging
pandas
partial-json-parser
pillow
pinecone-client
pinecone-plugin-inference
pinecone-plugin-interface
platformdirs
prometheus-fastapi-instrumentator
prometheus_client
propcache
proto-plus
protobuf
psutil
py-cpuinfo
py-spy
pyarrow
pyasn1
pyasn1_modules
pycountry
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
PyPDF2
python-dateutil
python-docx
python-dotenv
python-multipart
pytz
PyYAML
pyzmq
ray
redis
referencing
regex
requests
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
sentencepiece
sentry-sdk
setuptools
six
slowapi
smart-open
smmap
sniffio
starlette
streamlit
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
toml
torch
torchvision
tornado
tqdm
transformers
typing_extensions
tzdata
urllib3
uvicorn
virtualenv
watchdog
watchfiles
websockets
wrapt
yarl
zipp